+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](https://semver.org)
 * Screens/log: segmentation fault when toggling the log in a detached state.
 * Screens/log: scroll "jitter" with multi-line entries. #46
 * Screens/log: always behave in a "sliding window" manner. #20
+* Downloader: uncaught exception on missing `resolve()` plugin function.
 
 ### [v0.7.0] - 2019-02-28